# api/openapi.yaml
openapi: 3.0.3 # ignore `String does not match the pattern of "^3\.1\.\d+(-.+)?$".` warnings for now: https://github.com/oapi-codegen/oapi-codegen/issues/373
info:
  title: Noot API
  description: AI-powered nutrition logging API
  version: "1.0"
  contact:
    name: Grant Birkinbine
    url: https://github.com/noot-app
servers:
  - url: /api/v1
    description: API v1
paths:
  /consumption:
    post:
      summary: Log a consumption via audio
      description: Upload an audio file of spoken consumption data to be transcribed, parsed, and analyzed for nutrition information
      operationId: createConsumption
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                audio:
                  type: string
                  format: binary
                  description: Audio file (webm, opus, mp3, wav)
              required:
                - audio
      responses:
        "200":
          description: Successful consumption analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumptionResponse"
        "400":
          description: Bad request (invalid multipart form, no audio file, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /consumption/{id}:
    put:
      summary: Update a consumption record
      description: Update an existing consumption with edited nutrition data
      operationId: updateConsumption
      parameters:
        - name: id
          in: path
          required: true
          description: Consumption ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConsumptionRequest"
      responses:
        "200":
          description: Consumption updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumptionResponse"
        "400":
          description: Bad request (invalid request body, validation errors)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Consumption not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    
    delete:
      summary: Delete a consumption record
      description: Delete a consumption record by ID (used for redo functionality)
      operationId: deleteConsumption
      parameters:
        - name: id
          in: path
          required: true
          description: Consumption ID
          schema:
            type: string
      responses:
        "200":
          description: Consumption deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          description: Consumption not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /consumptions:
    get:
      summary: List consumptions
      description: Get a list of stored consumptions for the default user, optionally filtered by labels and date range. Supports tier-based date range limits - Free users can access up to 7 days, Pro users up to 365 days.
      operationId: getConsumptions
      parameters:
        - name: labels
          in: query
          description: Comma-separated list of label names to filter by
          schema:
            type: string
            example: "lunch,trigger-foods"
        - name: match
          in: query
          description: Label matching strategy - 'any' matches consumptions with at least one of the labels, 'all' matches consumptions with all labels
          schema:
            type: string
            enum: [any, all]
            default: any
        - name: start
          in: query
          description: Start date (RFC3339 format) - cannot be used together with 'days' parameter
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          description: End date (RFC3339 format) - cannot be used together with 'days' parameter
          schema:
            type: string
            format: date-time
        - name: days
          in: query
          description: Number of days to look back from today - cannot be used together with 'start/end' parameters. Free tier limited to 7 days, Pro tier to 365 days.
          schema:
            type: integer
            minimum: 1
            maximum: 365
      responses:
        "200":
          description: List of consumptions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumptionsResponse"
        "400":
          description: Invalid parameters (e.g., both start/end and days provided, invalid date format)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Access denied (subscription tier limitation - Free users limited to 7 days, Pro users to 365 days)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /goals:
    get:
      summary: Get nutrition goals
      description: Get resolved nutrition goals based on user profile (DRI) and any custom overrides
      operationId: getGoals
      parameters:
        - name: goal_name
          in: query
          required: false
          schema:
            type: string
          description: Name of the specific goal set to retrieve. If not provided, returns the active goal set.
      responses:
        "200":
          description: User nutrition goals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalsResponse"
        "404":
          description: User not found or goal set not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    
    put:
      summary: Update nutrition goals (Pro only)
      description: Set custom nutrition goal overrides for Pro users
      operationId: updateGoals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoalsRequest"
      responses:
        "200":
          description: Goals updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalsResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Access denied (Pro subscription required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /goals/sets:
    get:
      summary: List all goal sets (Pro only)
      description: Get a list of all custom goal sets for the user
      operationId: getGoalSets
      responses:
        "200":
          description: List of goal sets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalSetsResponse"
        "403":
          description: Access denied (Pro subscription required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /goals/sets/{name}:
    delete:
      summary: Delete a goal set (Pro only)
      description: Delete a specific custom goal set
      operationId: deleteGoalSet
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Name of the goal set to delete
      responses:
        "204":
          description: Goal set deleted successfully
        "400":
          description: Cannot delete the active goal set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Access denied (Pro subscription required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Goal set not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /goals/active:
    put:
      summary: Set active goal set (Pro only)
      description: Set which goal set is currently active for the user
      operationId: setActiveGoalSet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetActiveGoalRequest"
      responses:
        "200":
          description: Active goal set updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalsResponse"
        "403":
          description: Access denied (Pro subscription required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Goal set not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
        - code
        - timestamp
      properties:
        error:
          type: string
          description: Error message
        code:
          type: integer
          description: HTTP status code
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred
        stack:
          type: array
          items:
            type: string
          description: Stack trace (only in debug mode)
        trace_id:
          type: string
          description: Request trace ID
...
    ConsumptionResponse:
      type: object
      required:
        - id
        - transcript
        - items
        - summary
        - request_id
      properties:
        id:
          type: string
          description: Consumption ID (for future edits/deletes)
        transcript:
          type: string
          description: Transcribed text from audio
        labels:
          type: array
          items:
            $ref: "#/components/schemas/Label"
          description: Labels assigned to this consumption
        items:
          type: array
          items:
            $ref: "#/components/schemas/ItemWithNutrition"
          description: Items with complete nutrition information
        summary:
          $ref: "#/components/schemas/Summary"
        request_id:
          type: string
          description: Request identifier for tracking
...