# internal/api/gin.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Delete user biometrics
	// (DELETE /biometrics)
	DeleteUserBiometrics(c *gin.Context)
	// Get user biometrics
	// (GET /biometrics)
	GetUserBiometrics(c *gin.Context)
	// Update user biometrics
	// (PUT /biometrics)
	UpdateUserBiometrics(c *gin.Context)
	// Log a consumption via audio
	// (POST /consumption)
	CreateConsumption(c *gin.Context)
	// Get labels assigned to a consumption item
	// (GET /consumption-items/{id}/labels)
	GetConsumptionItemLabels(c *gin.Context, id string)
	// Assign labels to a consumption item
	// (POST /consumption-items/{id}/labels)
	AssignConsumptionItemLabels(c *gin.Context, id string)
	// Unassign a label from a consumption item
	// (DELETE /consumption-items/{id}/labels/{labelId})
	UnassignConsumptionItemLabel(c *gin.Context, id string, labelId string)
	// Delete a consumption record
	// (DELETE /consumption/{id})
	DeleteConsumption(c *gin.Context, id string)
	// Update a consumption record
	// (PUT /consumption/{id})
	UpdateConsumption(c *gin.Context, id string)
	// Get labels assigned to a consumption
	// (GET /consumption/{id}/labels)
	GetConsumptionLabels(c *gin.Context, id string)
	// Assign labels to a consumption
	// (POST /consumption/{id}/labels)
	AssignConsumptionLabels(c *gin.Context, id string)
	// Unassign a label from a consumption
	// (DELETE /consumption/{id}/labels/{labelId})
	UnassignConsumptionLabel(c *gin.Context, id string, labelId string)
	// List consumptions
	// (GET /consumptions)
	GetConsumptions(c *gin.Context, params GetConsumptionsParams)
	// Export nutrition data (Pro only)
	// (GET /export)
	ExportData(c *gin.Context, params ExportDataParams)
	// Get nutrition goals
	// (GET /goals)
	GetGoals(c *gin.Context, params GetGoalsParams)
	// Update nutrition goals (Pro only)
	// (PUT /goals)
	UpdateGoals(c *gin.Context)
	// Set active goal set (Pro only)
	// (PUT /goals/active)
	SetActiveGoalSet(c *gin.Context)
	// List all goal sets (Pro only)
	// (GET /goals/sets)
	GetGoalSets(c *gin.Context)
	// Delete a goal set (Pro only)
	// (DELETE /goals/sets/{name})
	DeleteGoalSet(c *gin.Context, name string)
	// Health check
	// (GET /health)
	GetHealth(c *gin.Context)
	// List labels
	// (GET /labels)
	GetLabels(c *gin.Context)
	// Create a new label
	// (POST /labels)
	CreateLabel(c *gin.Context)
	// Delete a label
	// (DELETE /labels/{id})
	DeleteLabel(c *gin.Context, id string)
	// Update a label
	// (PUT /labels/{id})
	UpdateLabel(c *gin.Context, id string)
	// Get nutrition trends
	// (GET /trends)
	GetTrends(c *gin.Context, params GetTrendsParams)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// DeleteUserBiometrics operation middleware
func (siw *ServerInterfaceWrapper) DeleteUserBiometrics(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.DeleteUserBiometrics(c)
}

// GetUserBiometrics operation middleware
func (siw *ServerInterfaceWrapper) GetUserBiometrics(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetUserBiometrics(c)
}

// UpdateUserBiometrics operation middleware
func (siw *ServerInterfaceWrapper) UpdateUserBiometrics(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UpdateUserBiometrics(c)
}

// CreateConsumption operation middleware
func (siw *ServerInterfaceWrapper) CreateConsumption(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.CreateConsumption(c)
}

// GetConsumptionItemLabels operation middleware
func (siw *ServerInterfaceWrapper) GetConsumptionItemLabels(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetConsumptionItemLabels(c, id)
}

// AssignConsumptionItemLabels operation middleware
func (siw *ServerInterfaceWrapper) AssignConsumptionItemLabels(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.AssignConsumptionItemLabels(c, id)
}

// UnassignConsumptionItemLabel operation middleware
func (siw *ServerInterfaceWrapper) UnassignConsumptionItemLabel(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Path parameter "labelId" -------------
	var labelId string

	err = runtime.BindStyledParameterWithOptions("simple", "labelId", c.Param("labelId"), &labelId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter labelId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UnassignConsumptionItemLabel(c, id, labelId)
}

// DeleteConsumption operation middleware
func (siw *ServerInterfaceWrapper) DeleteConsumption(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.DeleteConsumption(c, id)
}

// UpdateConsumption operation middleware
func (siw *ServerInterfaceWrapper) UpdateConsumption(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UpdateConsumption(c, id)
}

// GetConsumptionLabels operation middleware
func (siw *ServerInterfaceWrapper) GetConsumptionLabels(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetConsumptionLabels(c, id)
}

// AssignConsumptionLabels operation middleware
func (siw *ServerInterfaceWrapper) AssignConsumptionLabels(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.AssignConsumptionLabels(c, id)
}

// UnassignConsumptionLabel operation middleware
func (siw *ServerInterfaceWrapper) UnassignConsumptionLabel(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Path parameter "labelId" -------------
	var labelId string

	err = runtime.BindStyledParameterWithOptions("simple", "labelId", c.Param("labelId"), &labelId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter labelId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UnassignConsumptionLabel(c, id, labelId)
}

// GetConsumptions operation middleware
func (siw *ServerInterfaceWrapper) GetConsumptions(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetConsumptionsParams

	// ------------- Optional query parameter "labels" -------------

	err = runtime.BindQueryParameter("form", true, false, "labels", c.Request.URL.Query(), &params.Labels)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter labels: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "match" -------------

	err = runtime.BindQueryParameter("form", true, false, "match", c.Request.URL.Query(), &params.Match)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter match: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "start" -------------

	err = runtime.BindQueryParameter("form", true, false, "start", c.Request.URL.Query(), &params.Start)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter start: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "end" -------------

	err = runtime.BindQueryParameter("form", true, false, "end", c.Request.URL.Query(), &params.End)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter end: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "days" -------------

	err = runtime.BindQueryParameter("form", true, false, "days", c.Request.URL.Query(), &params.Days)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter days: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetConsumptions(c, params)
}

// ExportData operation middleware
func (siw *ServerInterfaceWrapper) ExportData(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ExportDataParams

	// ------------- Required query parameter "format" -------------

	if paramValue := c.Query("format"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument format is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "format", c.Request.URL.Query(), &params.Format)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter format: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "start" -------------

	err = runtime.BindQueryParameter("form", true, false, "start", c.Request.URL.Query(), &params.Start)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter start: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "end" -------------

	err = runtime.BindQueryParameter("form", true, false, "end", c.Request.URL.Query(), &params.End)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter end: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "metrics" -------------

	err = runtime.BindQueryParameter("form", true, false, "metrics", c.Request.URL.Query(), &params.Metrics)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter metrics: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ExportData(c, params)
}

// GetGoals operation middleware
func (siw *ServerInterfaceWrapper) GetGoals(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetGoalsParams

	// ------------- Optional query parameter "goal_name" -------------

	err = runtime.BindQueryParameter("form", true, false, "goal_name", c.Request.URL.Query(), &params.GoalName)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter goal_name: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetGoals(c, params)
}

// UpdateGoals operation middleware
func (siw *ServerInterfaceWrapper) UpdateGoals(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UpdateGoals(c)
}

// SetActiveGoalSet operation middleware
func (siw *ServerInterfaceWrapper) SetActiveGoalSet(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.SetActiveGoalSet(c)
}

// GetGoalSets operation middleware
func (siw *ServerInterfaceWrapper) GetGoalSets(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetGoalSets(c)
}

// DeleteGoalSet operation middleware
func (siw *ServerInterfaceWrapper) DeleteGoalSet(c *gin.Context) {

	var err error

	// ------------- Path parameter "name" -------------
	var name string

	err = runtime.BindStyledParameterWithOptions("simple", "name", c.Param("name"), &name, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter name: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.DeleteGoalSet(c, name)
}

// GetHealth operation middleware
func (siw *ServerInterfaceWrapper) GetHealth(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetHealth(c)
}

// GetLabels operation middleware
func (siw *ServerInterfaceWrapper) GetLabels(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetLabels(c)
}

// CreateLabel operation middleware
func (siw *ServerInterfaceWrapper) CreateLabel(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.CreateLabel(c)
}

// DeleteLabel operation middleware
func (siw *ServerInterfaceWrapper) DeleteLabel(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.DeleteLabel(c, id)
}

// UpdateLabel operation middleware
func (siw *ServerInterfaceWrapper) UpdateLabel(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UpdateLabel(c, id)
}

// GetTrends operation middleware
func (siw *ServerInterfaceWrapper) GetTrends(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetTrendsParams

	// ------------- Optional query parameter "metrics" -------------

	err = runtime.BindQueryParameter("form", true, false, "metrics", c.Request.URL.Query(), &params.Metrics)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter metrics: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "start" -------------

	err = runtime.BindQueryParameter("form", true, false, "start", c.Request.URL.Query(), &params.Start)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter start: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "end" -------------

	err = runtime.BindQueryParameter("form", true, false, "end", c.Request.URL.Query(), &params.End)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter end: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "days" -------------

	err = runtime.BindQueryParameter("form", true, false, "days", c.Request.URL.Query(), &params.Days)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter days: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetTrends(c, params)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.DELETE(options.BaseURL+"/biometrics", wrapper.DeleteUserBiometrics)
	router.GET(options.BaseURL+"/biometrics", wrapper.GetUserBiometrics)
	router.PUT(options.BaseURL+"/biometrics", wrapper.UpdateUserBiometrics)
	router.POST(options.BaseURL+"/consumption", wrapper.CreateConsumption)
	router.GET(options.BaseURL+"/consumption-items/:id/labels", wrapper.GetConsumptionItemLabels)
	router.POST(options.BaseURL+"/consumption-items/:id/labels", wrapper.AssignConsumptionItemLabels)
	router.DELETE(options.BaseURL+"/consumption-items/:id/labels/:labelId", wrapper.UnassignConsumptionItemLabel)
	router.DELETE(options.BaseURL+"/consumption/:id", wrapper.DeleteConsumption)
	router.PUT(options.BaseURL+"/consumption/:id", wrapper.UpdateConsumption)
	router.GET(options.BaseURL+"/consumption/:id/labels", wrapper.GetConsumptionLabels)
	router.POST(options.BaseURL+"/consumption/:id/labels", wrapper.AssignConsumptionLabels)
	router.DELETE(options.BaseURL+"/consumption/:id/labels/:labelId", wrapper.UnassignConsumptionLabel)
	router.GET(options.BaseURL+"/consumptions", wrapper.GetConsumptions)
	router.GET(options.BaseURL+"/export", wrapper.ExportData)
	router.GET(options.BaseURL+"/goals", wrapper.GetGoals)
	router.PUT(options.BaseURL+"/goals", wrapper.UpdateGoals)
	router.PUT(options.BaseURL+"/goals/active", wrapper.SetActiveGoalSet)
	router.GET(options.BaseURL+"/goals/sets", wrapper.GetGoalSets)
	router.DELETE(options.BaseURL+"/goals/sets/:name", wrapper.DeleteGoalSet)
	router.GET(options.BaseURL+"/health", wrapper.GetHealth)
	router.GET(options.BaseURL+"/labels", wrapper.GetLabels)
	router.POST(options.BaseURL+"/labels", wrapper.CreateLabel)
	router.DELETE(options.BaseURL+"/labels/:id", wrapper.DeleteLabel)
	router.PUT(options.BaseURL+"/labels/:id", wrapper.UpdateLabel)
	router.GET(options.BaseURL+"/trends", wrapper.GetTrends)
}